requirements.txt
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
black==25.1.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
isort==6.0.0
Mako==1.3.9
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
